Converted MouseX/MouseY in InputEvent to float, since the values don't represent pixels on screen and should not be INT

+++ src/common/engine/d_eventbase.h
@@ -70,8 +70,8 @@ struct FInputEvent
 	FString KeyString;
 	int KeyChar;
 	// for mouse
-	int MouseX;
-	int MouseY;
+	float MouseX;
+	float MouseY;
 
 	FInputEvent(const event_t *ev);
 };
